# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VandLabs Automobile Engine BFF
  version: 1.1.0
  description: >
    Proof-of-Engine HTTP contracts. These endpoints are implemented by the
    Next.js BFF today and can later proxy API Gateway/domain services without
    changing customer-facing UI contracts.
servers:
  - url: http://127.0.0.1:3000
paths:
...
  /api/leads:
    post:
      summary: Create a structured automotive lead
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateLead" }
      responses:
        "201":
          description: Lead accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  leadId: { type: string }
                  stage: { type: string, const: new }
                  vehicleId: { type: [string, "null"] }
                  message: { type: string }
        "400":
          description: Invalid lead payload
...
components:
  schemas:
...
    CreateLead:
      type: object
      required: [name, phone]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string, minLength: 1 }
        email: { type: string }
        intent:
          type: string
          enum: [enquiry, test_drive, finance, exchange]
        vehicleId: { type: string }
        notes: { type: string }
        whatsappConsent: { type: boolean }
        marketingConsent: { type: boolean }
        attribution:
          type: object
          additionalProperties: true
...